pyproject.toml
[project]
name = "lab01-part2-nusp"
version = "0.1.0"
description = "Pipeline de ingestão de dados containerizado"
authors = [
    { name = "Gustavo" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "deltalake>=1.5.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "polars>=1.39.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
